GTEST: Minor code refactoring related to CUDA context usage.
